Added better support for empty label string
There was an issue with using an empty label string where the summary label was at a different position than the recipient views causing a jump. The solution was to not base the summary label's constraint off of the to label.

Also, the constraints have been adjusted so that a space in the label is used for the padding instead of a constraint.